Add issue templates for reporting bugs etc.
The issues that have been added match what was already in markdown
files: report a bug and request a feature. These templates are based on
what I did for the temp-test-ioccc repo. Tested syntax and look and feel
in a personal repo for this purpose and it looks good (though maybe
slight improvements might be warranted).