Improve bug_report.sh for users w/o default gmake
Now has a new option -m make to allow one to set the path to make. This
is because if one is on a system that does not default to GNU make it
will likely fail. The script, before parsing args, tries to find gmake
and if that fails it will try and find make. If after parsing options
make cannot be found or is not a regular executable file it is
considered a command line error as without make this script is mostly
useless anyway and it simplifies the setting of exit codes.

The man page has been updated.

Changed BUG_REPORT_VERSION to "1.0.3 2024-06-26".