Change C standard to c17
It was stated that this would happen sometime in 2023 and we're almost
at July in 2024. One of the problem systems that required c11 will be
reaching EOL (thanks to Red Hat) on 30 June 2024 so it seems to not be a
problem. Rocky linux appears to work with c17 (although I only tested
the top level Makefile for that but it should work fine).