Fix limit_ioccc.sh not being generated at times
If one were to do make clobber iocccsize it would not build
soup/limit_ioccc.sh. It would, however, build the soup.a library, and
since it's a useful file to be built, and since tools rely on
soup.a, the soup.a rule now makes limit_ioccc.sh.

Doing it this way prevents having to add something like:

		${MAKE} -C soup limit_ioccc.sh

to all the rules in the Makefile and now one can do:

		make clobber iocccsize
		./test_ioccc/iocccsize_test.sh

and have it run okay. Without something like this (that can probably be
improved upon / cleaned up) one would have to build everything in order
for the file to be built (which maybe indicates a better procedure could
be devised).