revert commit 73a965d0e8d42b6f20a727524cdd65d417f1ab07
As a reminder, commit 73a965d0e8d42b6f20a727524cdd65d417f1ab07 was
an deliberate introduction of an error to test the effect of the
new `test.yml` file.

The net effect of `test.yml` is an advisory (not mandatory) check on the
result of a commit.

As a result of these successful tests, we now revert the deliberate
error and resume normal testing operations.