Update jparse/README.md and json_util_README.md
These files now state that the tools jfmt, jval and jnamval are VERY
INCOMPLETE and will be HEAVILY MODIFIED and will be almost entirely
rewritten. This is much quicker than updating the code and header files
to note this fact but it feels useful as they still exist.

When the tools are rewritten the notice will be removed from both files
and jparse/README.md will refer to them again, briefly giving an
overview (or at least referring to the other document) and how to render
the man pages.

This is partly done as I have updated the README.md in the jparse repo
itself to refer to the README.md here to give people a better idea,
noting that although one can certainly use the parser/library if they
wish, it is still tied to the mkiocccentry repo.