Bug fix txzchk -q option
Although it was quiet in some cases it still showed warnings. That was a
problem even if it wasn't for the fact that showing output with quiet
mode requested makes no sense.

An important point, however, is that if -w, -q will now be disabled.
That option is important for the test suite and that always applied but
since -q did not block warnings it was not detected.

I (re)discovered these problems (I had known about it previously but due
to the test suite needing warnings it was not fixed) when working on the
guidelines as I know txzchk best and there were some things that I felt
could be improved or better explained.

Updated TXZCHK_VERSION to "1.0.7 2024-07-10".
Updated TXZCHK_TEST_VERSION to "1.0.3 2024-07-10".

The txzchk_version in all the test json files have been updated (at
least those which are supposed to have the current version, which might
or might not be all of them).

Ran make release to rebuild parser (though that was not necessary as
jparse was not touched) and to test everything.