Removed -q from txzchk - now always shows warnings
At the request of Landon the '-q' option was removed from the tool as
the submit server should not have to specify -q to the tool to quiet
certain things.

The tool now always shows warnings so if one wants to have silent output
they can use 2>/dev/null. Previously this was controlled with the -q
switch. The -w option which was to show warnings even if -q was used
still exists but this should not normally be required to get warning
messages. The use of it simply sets warn_output_allowed to true and
msg_warn_silent to false.

The punny messages in txzchk(1) are now in their own option (or actually
options): -e and -f; -f requires an argument that can be converted to an
unsigned integer but you will have to either try it out, read the source or
check the man page if you wish to know more as I (@xexyl) do not want to spoil
the fun of finding it (including the additional ones)! mkiocccentry(1) will
use the first option but not the second one. This was suggested on GitHub just
like the removal of -q was.

Updated TXZCHK_VERSION to "1.0.8 2024-07-11".

Ran make rebuild_txzchk_test_errors (due to slight message changes) and then
make release to test everything.