Add -e switch to mkiocccentry for txzchk
The mkiocccentry_test.sh script always uses this option to simplify
things and because the ones who run it most often (or make prep or make
release, the latter of which was run) appreciate the option as well (and
it's quicker to do which allows getting to other more important things).

MKIOCCCENTRY_VERSION changed to "1.0.7 2024-07-12".

Man page mkiocccentry.1 updated.