Add many valid and invalid test JSON files
With MUCH GRATITUDE (THANK YOU!) to the JSONTestSuite repo
(https://github.com/nst/JSONTestSuite) maintainers we have added many
good and bad JSON files of more unique kinds (as in kinds that some
parsers get wrong, both false positive and false negative) to the
jparse/test_jparse/test_JSON subdirectories (good/ and bad/) and all is
good.

A note is that this does slow down make test, make prep (and make
slow_prep) a bit but this is worth it because we can now be even more
sure that everything is in order if by chance an error is introduced
into the parser. We did not (by an oversight) document that new GitHub
workflows have been added to run tests on commit and push that will show
any problems before a commit is merged to the master branch (well if
Landon does a push that is another matter entirely) so that is another
recent safety mechanism.

The jparse/README.md file has been updated: moved the history section to
the very bottom and added the section on the testing procedure (I also
slightly amended the heading of the additional tools that are VERY
INCOMPLETE and will very likely be rewritten almost entirely) with a
warning so it stands out even more.