Update CHANGES.md
Add that the make test, make prep and make slow_prep rules now have (as
of the last commit) a timestamp at beginning and end of the rules (this
is very helpful at times).

I also added that we recently (thanks @SirWumpus!) added a workflow so
that we can detect a problem before (unless Landon does the commit of
course :-) ) a commit is merged into master.