Many updates and fixes to FAQ
Added version to FAQ - first version is 28.0.0 2024-07-31 (same format
to the temp-test-ioccc repo but since the version is just being added it
is at 28.0.0).

Removed FAQs about how to use the tools in favour of just one FAQ about
how to participate in the contest. This FAQ links to the proper FAQs in
the temp-test-ioccc repo (the rendered html) with the note that the
links will have to be updated once the great fork merge happens (this is
in XXX html comments), the rules and the guidelines. This way the FAQ
here focuses on how to obtain, compile and get help on these tools (as
well as if necessary report bugs though that is also in the other repo's
FAQ). Also the information is not duplicated.

It is however unclear to me if the FAQ in the other repo should just
refer to the FAQ here: maybe (probably I would guess) it should but
right now it does not. This can be decided and fixed later.

Reorganised the order of the questions in what I think flows a bit
better (for certain definitions of better I suppose).

Added new named tags/links and the table of contents links to those
instead of something like '#0-where-can-i-find-the-mkiocccentry-tool'
which would have to be updated if the question changes or even if the
number changes (which as this commit shows can happen).