Add JSON reference files for chkentry
This commit is to begin the process of updating chkentry(1) to validate
.entry.json and author_handle.json files. The reference files use Landon
as the winner/author as not only do I not want to be conceited :-) but
he's a judge and it works well as a reference. A few notes of interest
here.

The author_handle.reference.json file has the current email address of
the Judges (which changes on a yearly basis). It was not clear if I
should have null instead but this can be decided later.

Where needed the entry_ids and author_handles etc. match in the two
files.

More can be done with this as more is discussed. The other data files
that will be needed for this hacking of chkentry can be added later as
more of the issue is discussed but this is a nice starting point.