Add struct entry and struct author_handle
These WILL correspond to the .entry.json and the author_handle.json
files in the other repo for chkentry (LATER). SOME (not all) test_foo()
functions have been added and the ones that act on the entry_id and
directory and in hindsight they will have to work together, perhaps even
a new function that is given both strings (to verify that the YYYY in
YYYY_winner matches the YYYY/winner, for instance).

The chkentry has NOT been updated to use these functions as they (and
especially object2entry()) have not been completed and at least one is
missing too (object2author_handle or something like that). It is hoped
that starting this will help discussing what is needed.

Perhaps the most important part of these changes are the new structs but
some of the code that goes along with it has been started, even if in
need of modification, especially the object2foo functions. Going back it
might have been better if just the structs were added but as I started the
others I might as well get them in too.

In other words some of this MIGHT HAVE BEEN PREMATURE but as discussion
briefly started I have done these (it might have been better to just
add the structs but as I wanted to get a bit in I did more, in other
words).

XXX comments have been added accordingly.

New defined macros: for versions of the json files (can be found in the
files in the other repo) and for min/max of directory length (based on
the max handle length - + 5 for the YYYY plus the '/' and '_').