remove jfmt, jval, jnamval
We remove jfmt, jval, and jnamval from the code base until they are ready
to be tested so that the general public and/or IOCCC submitters do not
become confused by these "not ready" tools and file bug reports on them.

In [GH-issuecomment-2282221142](https://github.com/ioccc-src/mkiocccentry/issues/523#issuecomment-2282221142)
a model based on the [JSONPath.sh](issuecomment-2282221142) command line
be considered for jfmt, jval, jnamval.

In
[GH-issuecomment-2278613143](https://github.com/ioccc-src/mkiocccentry/issues/928#issuecomment-2278613143)
it has been stated that jfmt, jval, jnamval are "_incomplete and will
be heavily modified and almost entirely rewritten_".  In light of the
the [JSONPath.sh](issuecomment-2282221142) command line model, the
"_entirely rewritten_" is much more likely.

Therefore we remove nearly all references to jfmt, jval, and jnamval.

We have **NOT** removed every reference to jfmt, jval, and jnamval:

- The `CHANGES.md` file will continue to contain historical references
to jfmt, jval, and jnamval.

- The `make uninstall` will continue to remove any installed binary and
man page related to jfmt, jval, and jnamval.

- The `jparse/.gitignore` file ignores the jfmt, jval, and jnamval binary
files to make it easier for future developers to work on those tools.

- The _JSON_UTIL_MATCH_TYPE_ types are left in `jparse/json_util.h`
as they related to types of JSON elements that could be matched later on.

- The `struct json_util_number_range` and `struct json_util_number`
files to make it easier for future developers to work on those tools.

- The `json_util_parse_number_range()` function in `jparse/json_util.c`
is left because it may be useful for future tools need to compare numbers.

- The `json_util_number_in_range` function in `jparse/json_util.c`
is left because it may be useful for future tools need to compare numbers.

- The `json_util_parse_st_level_option()` function in `jparse/json_util.c`
is left because it may be useful for future tools need to focus on parse
tree levels.

Performed `make release` to test the above under macOS.