update CHANGES for Release 1.4.2 2024-08-21
NOTE: Release 1.4.2 2024-08-21 is not finished, we are
just updating the CHANGES file.

Formatting a most of the `test_ioccc/test_JSON/auth.json/good/*.json`
and most of the `test_ioccc/test_JSON/info.json/good/*.json` files.
We did **NOT** modify bad files, and did **NOT** modify the
good files that used a minimum of whitespace.

NOTE: We have NOT added new test cases, just canonically formatted
      most good existing test cases.

NOTE: We have NOT updated `MKIOCCCENTRY_VERSION`.