remove include of 1337.mk from Makefile.example
Removed the following lines at the end of `Makefile.example`:

```makefile

-include 1337.mk ../1337.mk ../../1337.mk
```

Those lines are not needed by people forming IOCCC submissions
and only adds complexity to their tasks.

The IOCCC judges can, if needed, easily append such lines
to a submission's Makefile when the need to do so.

The IOCCC judges will need to integrate use of `var.mk`,
replacing the "common tool locations" set with an
include of that Makefile anyway, so they might as well
append the above include lines while they are at it.