Add a number of test files to test_JSON - #931
Added a number of test JSON files to demonstrate that the JSON parser is
okay with mixed spacing (spaces and tabs or different number of spaces
with or without tabs etc.), no spacing and so on, even though the
mkiocccentry(1) tool will form json files with multiples of 4 spaces and
no tabs.

Updated MKIOCCCENTRY_VERSION to "1.0.8 2024-08-23".