Move open_json_dir_file to open_dir_file (util.c)
As this function is useful for more than just chkentry.c and more than
just JSON (though this is all it's used for currently) it has been
renamed and moved to jparse/util.c.

This commit is, although useful, not critical at this time. No versions
have changed because it is not clear yet what version strings should be
changed and because this might not (yet) warrant a new release.