Fix hostchk.sh
With the recent updates to the inclusion of the different repos and the
MKIOCCCENTRY_SRC a problem in linux for hostchk.sh had to be resolved.
This might also be a problem in macOS but since the libraries were
already installed this is not detected (it seems that in linux if the
other libraries were installed this also was not a problem).

There still exists an issue with building, however, as the make
dependencies have system paths (cc -MM will solve this so it might be
the way independ works but this is not clear to me at this time).
Nonetheless the hostchk.sh should at least now work.

New version of hostchk.sh is 1.0.2 2024-08-28.