Rename dbg.a and dyn_array.a for outside this repo
Rename dbg.a and dyn_array.a to libdbg.a and libdyn_array.a
respectively. These respective repos have had the respective change done as
well. This is important for the jparse repo (not the jparse/ subdirectory here
though this has also been updated) to be usable outside this repo without
having to have ../dbg and ../dyn_array directories containing the correct
respective libraries. The jparse repo (https://github.com/xexyl/jparse)
still has to be updated (and there might be some additional changes here
as well) but that cannot happen until the dbg and dyn_array repos have
also had this done (or rather merged).

The Makefile does NOT use -ldbg or -ldyn_array (yet?).

Updated MKIOCCCENTRY_REPO_VERSION to match this release version (this means it
jumped from 1.5.2 to 1.5.4 as it was not updated when 1.5.2 2024-08-28
was updated to 1.5.3 2024-08-29).

For a little better details on this commit see the CHANGES.md file
(search for the line starting with '## Release 1.5.4 2024-08-29'
in the file itself or if viewed on GitHub the line that starts with
'Release 1.5.4 2024-08-29').