Rename library filenames for external repos
Rename dbg.a, dyn_array.a and jparse.a to libdbg.a, libdyn_array.a and
libjparse.a, respectively. These respective repos have had the
respective change done as well (well the jparse one has not been
committed yet as it needs the other libraries in order to work).

This is important for the jparse repo (not the jparse/ subdirectory here
though this has also been updated) to be usable outside this repo
without having to have ../dbg and ../dyn_array directories containing
the correct respective libraries. The jparse repo
(https://github.com/xexyl/jparse) still has to be updated (and there
might be some additional changes here as well) but that cannot happen
until the dbg and dyn_array repos have also had this done (or rather
merged).

The Makefile does NOT use -ldbg, -ldyn_array or -ljparse (yet?).
Obviously if it does it'll use the -I option.

Updated MKIOCCCENTRY_REPO_VERSION to match this release version (this
means it jumped from 1.5.2 to 1.5.4 as it was not updated when 1.5.2
2024-08-28 was updated to 1.5.3 2024-08-29).

Ran make prep to test these changes.