update dbg, dyn_array, jparse and .gitignore
We update `dbg/` from the dbg repo.

We prevent files with `/usr/include` and with `/usr/local/include`
from showing up in the dependency list (via `make depend`).

The old `dbg.a` is now a symlink to `libdbg.a`.

We update `dyn_array/` from the dyn_array repo.

We prevent files with `/usr/include` and with `/usr/local/include`
from showing up in the dependency list (via `make depend`).

The old `dyn_array.a` is now a symlink to `libdyn_array.a`.

We update `jparse/` from the jparse repo.

We update `.gitignore` to ignore "build.log", "jparse.clone.xexyl"
and "jparse.clone.lcn2".

Performed `make release` to test the above under macOS.