Add to dyn_array/Makefile -L../dbg
The way the dyn_array repo Makefile now work (required for dbg for
jparse repo) is that it links in the dbg api by -ldbg (hence renaming it
to libdbg.a). But since GitHub does not have it installed this fails.
This suggests that compiling this repo might also fail if a person does
not have it installed. Thus I added -L../dbg. If this works then the
dyn_array repo's Makefile will have to have this added too.