Sync jparse subdirectory from jparse repo
With recent updates to the jparse repo
(https://github.com/xexyl/jparse) there was a need to sync that repo to
the jparse/ subdirectory. No code changes were made and the only
functionality changes made are that a few more header files are
installed (and now into a subdirectory - /usr/local/include/jparse) and
an uninstall rule was added (for those who wish to deobfuscate their
system a bit :-) ). The uninstall rule in the top level Makefile (this
repo) does not run make uninstall in other directories but perhaps it
should.

Did the following steps to make sure things are okay:

    # to be sure the tree is clean
    git status
    # to see if things can still build and test
    make release
    # to be sure the tree is clean
    make clobber

    make jparse.help
    make jparse.recreate_clone
    make jparse.update_from_clone

    # to see if anything changed
    git status
    # to see if things can still build and test
    make release

.. and all is good.

Updated the repo release version to "1.5.6 2024-08-31".