Attempt to workaround path in ioccc_test.sh
Due to jparse being in its own repo an issue in the execution of
jparse/test_jparse/jparse_test.sh on test_ioccc/test_JSON/general.json
is not working in the workflow on GitHub even though it does work
locally. To try and work past this I have used -Z test_ioccc and then
../jparse for that which is under the jparse directory. This also works
locally and hopefully it will work with GitHub workflows.

Also I have committed the changelog which was supposed to be committed
earlier on.