Add general.json/bad_loc files
This should resolve the GitHub workflow problem not working when it
works locally. The error (upon closer inspection) was that the directory
could not be found (which was necessary as the option to check bad
locations file was removed in favour of always testing it as if it does
not work it indicates a problem). This makes sense because you can't add
empty files in git but I had the directory locally hence why make test
worked for me.

If this works then one final thing for the script has to be done.