Add {dbg,dyn_array}/.github to .gitignore
The reason that the jparse/.github did not seem to be deleted in the
commit is not what does not make sense (even if rm -r is used it should
have been registered as being removed and I would have seen this in git
status) in the previous commit but rather because I (or Landon maybe)
had already put /jparse/.github in the .gitignore.

Now the .gitignore file includes /dbg/.github and /dyn_array/.github as
well as /dbg/dbg_test and /dbg/dbg_example.