Sync jparse/ from repo
Some fixes to the jparse Makefiles had to be made and these are now
synced to this repo. These fixes include:

- Fix make clobber to remove jparse_test.log and Makefile.orig.

- Fix make legacy_clobber to remove jparse.a.

- Fix ${RM} in Makefiles to use ${Q} variable (not in make depend as it is
used in an earlier command in the multiple line commands), in some cases changed
from the wrong variable, and ${RM_V} (where this was not done).

- Do not use -r in rm in Makefiles unless removing a directory, for safety.

The MKIOCCCENTRY_REPO_VERSION was updated to reflect the CHANGES.md.