Fix Eszett in utf8_posix_map.c to be 'ss'
The Eszett ('ß'), also known as 'sharp S' (what it was labelled as,
though in all caps), should be a double 's' so it now is (the table does
allow for multiple letters e.g. for 'Æ' -> 'AE'). In parentheses there
is now the word ESZETT which can help find it (for those of us used to
the German term, a language that uses it e.g. in Straße).

The repo release version string in soup/version.h has been updated to
reflect this change.

CHANGES.md has been updated for this update as well.