Bug fixes and enhancements in jstr{de,en}code
Synced jparse/ from the jparse repo (https://github.com/xexyl/jparse/).
This makes some important fixes to the tools jstrdecode(1) and
jstrencode(1) where in jstrdecode(1) the -Q option did not work
and for both the printing of everything should happen after everything
is parsed (especially for the -Q option but not strictly for that
reason). The pr_jparse_test tool had a bug fix as well where the -h
option did not work.