Fix make check_man, sync jparse
Fix check_man Makefile rule in all Makefiles. Previously it ignored the result
which meant that prep.sh never reported a problem in a problematic man page.

The dbg and dyn_array repos need to have this fix applied and I will do
this as well.

Synced jparse repo for recent changes, which includes the above fix, a
man page fix and a new helper function besides.