Sync jparse/ from jparse repo
Sync jparse/ from jparse repo (https://github.com/xexyl/jparse/). The
following was added to the jparse CHANGES.md file:

- Implement boolean unicode of struct json_string in the decoding functions.
If json_conv_string() finds that calling json_decode() which calls
decode_json_string() causes the bool unicode to be false it sets the
converted boolean to false. This will then flag an error in parsing if there
is an invalid Unicode symbol. This does not mean that decoding all symbols work:
it simply means that we detect if there are invalid Unicode symbols. The check
is done on the original string but if it turns out it has to be done on the
decoded string that can be done easily.

- The json_decode() had a bug fix: it allocated memory for the return string
when decode_json_string() does that and this caused a memory leak.

- If decode_json_string() (which is now a static function in json_parse.c)
detects an error, the allocated memory is freed before returning NULL.

- Updated JSON_PARSER_VERSION to "1.1.6 2024-09-24".

- Updated JPARSE_REPO_VERSION to "1.0.13 2024-09-24".

- Rename jenc to byte2asciistr in json_parse.c to avoid confusion about its
purpose.

- Expand the output of jstrencode -t and jstrdecode -t to express that the
encode/decode tests have not yet been written. This depends on bug #13 being
resolved first.

- Changed optimisation flags in the Makefiles to not specify -g3 as debug
symbols are almost useless when optimising and we have -O3 in use. During
debugging one can always use:

    make C_OPT="-g3" clobber all

to compile in debug symbols or have a file in the respective directories (those
needed) called makefile.local with the line:

    C_OPT= -g3

which is used for development purposes but should not normally be done.

- Fix potential use without initialisation of inputlen in jstrencode.c.

- Add function decode_json_string() to help simplify the json_decode()
function as it's quite long. This new function takes the length and calculated