Sync jparse from jparse repo
This includes a new JSON file with some details about it as well as some
fixes in the README.md file (newer synopsis of some tools) as well as a
couple functions added.

A typo fix was made in soup/is_available.sh which has to be done in
soup/ as well.