Update README.md files - #955
The top level README.md has been greatly simplified by removing extra
details. It now only talks about using the repo to package a submission
to the IOCCC with some links to the FAQ.

The new soup/README.md file has information for those who wish to
develop for this repo: in particular requirements.

The new test_ioccc/README.md file has much of what used to be in the top
level README.md file. This includes a list of the tools in the repo, or
at least the main ones. It is possible this belongs in soup/README.md
instead.

The FAQ has been updated with some formatting fixes.