Sync jparse from jparse repo
This syncing should resolve JSON decoding bugs. Although a couple things
have to be done still to close that issue
(https://github.com/xexyl/jparse/issues/13) out it appears everything is
good now. Some 'good' JSON files were changed to be 'bad' as they have
invalid UTF-8 bytes, it seems, but if this turns out to be false (or if
the so-called JSON spec allows for accepting it), this can be changed
back, especially as some test suites suggest that those files are legal,
even though multiple sources say otherwise.

Rather than duplicate the rest of the changes here or in the CHANGES.md
file, I refer anyone who must know to the jparse CHANGES.md file or its
git log.